feat(vsc): hide login in case of spfx project
